// src/api-client/model/rentalStatusUpdateRequestRentalStatus.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */

/**
 * 변경할 대여 상태
 */
export type RentalStatusUpdateRequestRentalStatus = typeof RentalStatusUpdateRequestRentalStatus[keyof typeof RentalStatusUpdateRequestRentalStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RentalStatusUpdateRequestRentalStatus = {
  PENDING: 'PENDING',
  CANCEL: 'CANCEL',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
  RENTAL: 'RENTAL',
  RETURN_PENDING: 'RETURN_PENDING',
  RETURN_CONFIRMED: 'RETURN_CONFIRMED',
  RETURNED: 'RETURNED',
} as const;